feat: enhance post creation with author, likes, and comments fields; update error message for avatar requirement

+++ 4-Blog-Api/src/utils/cloudinary.js
@@ -16,7 +16,8 @@ const uploadOnCloudinary = async(localFilePath) => {
 		fs.unlinkSync(localFilePath);
 		return response;
 	} catch (error) {
-		fs.unlinkSync(localFilePath);
+		console.log(error);
+		if(fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
 		return null;
 	}
 }
